+++ src/network/client.rs
@@ -7,8 +7,8 @@ use crate::network::future::Future;
 use crate::network::handler::{ConnectionError, Credentials};
 use crate::network::protocol::{Protocol, Resp3};
 use crate::network::timeout::{Timeout, TimeoutError};
-use crate::subscribe::client::{Error, Subscription};
-use crate::subscribe::messages::ToPushMessage;
+use crate::subscription::client::{Error, Subscription};
+use crate::subscription::messages::ToPushMessage;
 use alloc::string::String;
 use bytes::Bytes;
 use core::fmt::{Debug, Formatter};
